system_kafka_consumers_improvements: split test
